+++ probe-rs/src/probe/stlink/mod.rs
@@ -38,6 +38,7 @@ const STLINK_MAX_WRITE_LEN: usize = 0xFFFC;
 #[derive(Debug)]
 pub struct STLink<D: StLinkUsb> {
     device: D,
+    name: String,
     hw_version: u8,
     jtag_version: u8,
     protocol: WireProtocol,
@@ -53,8 +54,10 @@ impl DebugProbe for STLink<STLinkUSBDevice> {
     fn new_from_selector(
         selector: impl Into<DebugProbeSelector>,
     ) -> Result<Box<Self>, DebugProbeError> {
+        let device = STLinkUSBDevice::new_from_selector(selector)?;
         let mut stlink = Self {
-            device: STLinkUSBDevice::new_from_selector(selector)?,
+            name: format!("ST-Link {}", &device.info.version_name),
+            device,
             hw_version: 0,
             jtag_version: 0,
             protocol: WireProtocol::Swd,
@@ -71,7 +74,7 @@ impl DebugProbe for STLink<STLinkUSBDevice> {
     }
 
     fn get_name(&self) -> &str {
-        "ST-Link"
+        &self.name
     }
 
     fn speed(&self) -> u32 {
@@ -147,6 +150,17 @@ impl DebugProbe for STLink<STLinkUSBDevice> {
                 commands::JTAG_ENTER_SWD
             }
         };
+
+        // Check and report the target voltage.
+        let target_voltage = self.get_target_voltage()?.expect("The ST-Link returned None when it should only be able to return Some(f32) or an error. Please report this bug!");
+        if target_voltage < crate::probe::LOW_TARGET_VOLTAGE_WARNING_THRESHOLD {
+            log::warn!(
+                "Target voltage (VAPP) is {:2.2} V. Is your target device powered?",
+                target_voltage
+            );
+        } else {
+            log::info!("Target voltage (VAPP): {:2.2} V", target_voltage);
+        }
 
         let mut buf = [0; 2];
         self.send_jtag_command(
@@ -242,16 +256,42 @@ impl DebugProbe for STLink<STLinkUSBDevice> {
         Some(self as _)
     }
 
-    fn get_arm_interface<'probe>(
-        self: Box<Self>,
-    ) -> Result<Option<Box<dyn ArmProbeInterface + 'probe>>, DebugProbeError> {
-        let interface = StlinkArmDebug::new(self)?;
-
-        Ok(Some(Box::new(interface)))
-    }
-
     fn has_arm_interface(&self) -> bool {
         true
+    }
+
+    fn into_probe(self: Box<Self>) -> Box<dyn DebugProbe> {
+        self
+    }
+
+    fn try_get_arm_interface<'probe>(
+        self: Box<Self>,
+    ) -> Result<Box<dyn ArmProbeInterface + 'probe>, (Box<dyn DebugProbe>, DebugProbeError)> {
+        match StlinkArmDebug::new(self) {
+            Ok(interface) => Ok(Box::new(interface)),
+            Err((probe, err)) => Err((probe.into_probe(), err)),
+        }
+    }
+
+    fn get_target_voltage(&mut self) -> Result<Option<f32>, DebugProbeError> {
+        let mut buf = [0; 8];
+        match self
+            .device
+            .write(&[commands::GET_TARGET_VOLTAGE], &[], &mut buf, TIMEOUT)
+        {
+            Ok(_) => {
+                // The next two unwraps are safe!
+                let a0 = (&buf[0..4]).pread_with::<u32>(0, LE).unwrap() as f32;
+                let a1 = (&buf[4..8]).pread_with::<u32>(0, LE).unwrap() as f32;
+                if a0 != 0.0 {
+                    Ok(Some((2.0 * a1 * 1.2 / a0) as f32))
+                } else {
+                    // Should never happen
+                    Err(StlinkError::VoltageDivisionByZero.into())
+                }
+            }
+            Err(e) => Err(e),
+        }
     }
 }
 
@@ -315,22 +355,6 @@ impl DAPAccess for STLink<STLinkUSBDevice> {
             Err(StlinkError::BlanksNotAllowedOnDPRegister.into())
         }
     }
-
-    fn into_probe(self: Box<Self>) -> Box<dyn DebugProbe> {
-        self
-    }
-}
-
-impl<'a> AsRef<dyn DebugProbe + 'a> for STLink<STLinkUSBDevice> {
-    fn as_ref(&self) -> &(dyn DebugProbe + 'a) {
-        self
-    }
-}
-
-impl<'a> AsMut<dyn DebugProbe + 'a> for STLink<STLinkUSBDevice> {
-    fn as_mut(&mut self) -> &mut (dyn DebugProbe + 'a) {
-        self
-    }
 }
 
 impl<D: StLinkUsb> Drop for STLink<D> {
@@ -359,29 +383,6 @@ impl<D: StLinkUsb> STLink<D> {
 
     /// Firmware version that adds multiple AP support.
     const MIN_JTAG_VERSION_MULTI_AP: u8 = 28;
-
-    /// Reads the target voltage.
-    /// For the china fake variants this will always read a nonzero value!
-    pub fn get_target_voltage(&mut self) -> Result<f32, DebugProbeError> {
-        let mut buf = [0; 8];
-        match self
-            .device
-            .write(&[commands::GET_TARGET_VOLTAGE], &[], &mut buf, TIMEOUT)
-        {
-            Ok(_) => {
-                // The next two unwraps are safe!
-                let a0 = (&buf[0..4]).pread_with::<u32>(0, LE).unwrap() as f32;
-                let a1 = (&buf[4..8]).pread_with::<u32>(0, LE).unwrap() as f32;
-                if a0 != 0.0 {
-                    Ok((2.0 * a1 * 1.2 / a0) as f32)
-                } else {
-                    // Should never happen
-                    Err(StlinkError::VoltageDivisionByZero.into())
-                }
-            }
-            Err(e) => Err(e),
-        }
-    }
 
     /// Get the current mode of the ST-Link
     fn get_current_mode(&mut self) -> Result<Mode, DebugProbeError> {
@@ -522,7 +523,7 @@ impl<D: StLinkUsb> STLink<D> {
             self.jtag_speed_khz = current;
         }
 
-        self.get_target_voltage().map(|_| ())
+        Ok(())
     }
 
     /// sets the SWD frequency.
@@ -1119,7 +1120,9 @@ struct StlinkArmDebug {
 }
 
 impl StlinkArmDebug {
-    fn new(probe: Box<STLink<STLinkUSBDevice>>) -> Result<Self, DebugProbeError> {
+    fn new(
+        probe: Box<STLink<STLinkUSBDevice>>,
+    ) -> Result<Self, (Box<STLink<STLinkUSBDevice>>, DebugProbeError)> {
         let state = ArmCommunicationInterfaceState::new();
 
         // Determine the number and type of available APs.
@@ -1127,7 +1130,10 @@ impl StlinkArmDebug {
         let mut interface = Self { probe, state };
 
         for ap in valid_access_ports(&mut interface) {
-            let ap_state = interface.read_ap_information(ap)?;
+            let ap_state = match interface.read_ap_information(ap) {
+                Ok(state) => state,
+                Err(e) => return Err((interface.probe, e)),
+            };
 
             log::debug!("AP {}: {:?}", ap.port_number(), ap_state);
 
@@ -1353,6 +1359,12 @@ impl<'probe> ArmProbeInterface for StlinkArmDebug {
         self.state.ap_information.len()
     }
 
+    fn target_reset_deassert(&mut self) -> Result<(), ProbeRsError> {
+        self.probe.target_reset_deassert()?;
+
+        Ok(())
+    }
+
     fn close(self: Box<Self>) -> Probe {
         Probe::from_attached_probe(self.probe)
     }
@@ -1454,18 +1466,6 @@ where
             values,
         )?;
         Ok(())
-    }
-}
-
-impl<'a> AsRef<dyn DebugProbe + 'a> for StlinkArmDebug {
-    fn as_ref(&self) -> &(dyn DebugProbe + 'a) {
-        self.probe.as_ref()
-    }
-}
-
-impl<'a> AsMut<dyn DebugProbe + 'a> for StlinkArmDebug {
-    fn as_mut(&mut self) -> &mut (dyn DebugProbe + 'a) {
-        self.probe.as_mut()
     }
 }
 
@@ -1694,6 +1694,7 @@ mod test {
         fn build(self) -> STLink<MockUsb> {
             STLink {
                 device: self,
+                name: "Mock STlink".into(),
                 hw_version: 0,
                 protocol: WireProtocol::Swd,
                 jtag_version: 0,